requirements/base.txt
Django==4.2.7
asgiref==3.8.1
sqlparse==0.5.3
djangorestframework==3.15.2
django-filter==23.5
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pillow==10.3.0
python-dotenv==1.0.0
django-extensions==4.1.0
